Working on docker build
- add gcc-13 installation to docker file

Dockerfile is currently broken, because protoc is not discovered
correctly for generating GRPC stubs.

+++ docker/build-conan-env.sh
@@ -1,18 +1,21 @@
 #!/usr/bin/env bash
 cache=""
 image_name="network_components/conan-env:1.0"
+skip_conan=false
 
 usage() {
-  echo "Usage: $0 [-r|--no-cache]"
+  echo "Usage: $0 [-r|--no-cache] [-s|--skip-conan"
   echo
   echo "Options:"
   echo "  -r, --no-cache   Rebuilds the whole container, resetting cached layers."
   echo "                   (see docker build --no-cache)"
+  echo "  -s, --skip-conan Omits extracting conan information into local conan folder."
 }
 
 while [[ "$#" -gt 0 ]]; do
   case "$1" in
     -r|--no-cache) cache="--no-cache" ;;
+    -s|--skip-conan) skip_conan=true ;;
     -h|--help) usage
       exit 0 ;;
     *) echo "Invalid argument: $1" >&2
@@ -28,6 +31,11 @@ docker build \
    `#--progress=plain` \
    -t "${image_name}" \
    -f Dockerfile.conan-env .
+
+if [[ "$skip_conan" == true ]]; then
+  echo "Skipping local conan preparation."
+  exit 0;
+fi
 
 echo "Prepare local conan folder ..."
 docker run --rm --entrypoint="printenv" "${image_name}" "CONAN_USER_HOME"
